修正人员定位实时监测因数据表结构调整导致的各种问题

+++ src/main/java/cn/ccrise/spimp/location/entity/LocationArea.java
@@ -55,7 +55,7 @@ public class LocationArea {
 	/**
 	 * 区域内当前人数
 	 */
-	private Integer curPersonNum;
+	private Long curPersonNum;
 
 	/**
 	 * 状态
@@ -97,7 +97,7 @@ public class LocationArea {
 	}
 
 	@Transient
-	public Integer getCurPersonNum() {
+	public Long getCurPersonNum() {
 		return curPersonNum;
 	}
 
@@ -153,7 +153,7 @@ public class LocationArea {
 		this.createTime = createTime;
 	}
 
-	public void setCurPersonNum(Integer curPersonNum) {
+	public void setCurPersonNum(Long curPersonNum) {
 		this.curPersonNum = curPersonNum;
 	}
